+++ src/main/java/fr/insa/optimod/modele/Noeud.java
@@ -3,12 +3,12 @@ package fr.insa.optimod.modele;
 import java.util.ArrayList;
 
 public class Noeud {
-    protected int id;
+    protected Long id;
     protected Double longitude;
     protected Double latitude;
     protected ArrayList<Troncon> adjacense;
 
-    public Noeud(int id, Double longitude, Double latitude) {
+    public Noeud(Long id, Double longitude, Double latitude) {
         this.id = id;
         this.longitude = longitude;
         this.latitude = latitude;
@@ -24,7 +24,7 @@ public class Noeud {
         return longitude;
     }
 
-    public int getId() {
+    public Long getId() {
         return id;
     }
